updated changes

+++ public/javascripts/delete_user.js
@@ -2,9 +2,11 @@ $(clickHandler);
 
 function clickHandler() {
   $(".deleteButton").click(function() {
+    if (confirm('Are you sure you want to delete?')) {
     const uid = $(this).attr("data-uid");
     console.log("delete: ", uid);
     deleteUser(uid);
+    } else{}
   });
 }
